refactor: move code around to better places
